# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"WalletType": {
				"enum": [
					"spot",
					"funding",
					"isolated_margin",
					"cross_margin",
					"usd_m_futures",
					"usdt_m_futures",
					"usdc_m_futures",
					"coin_m_futures",
					"coin_m_swap",
					"options",
					"earn",
					"derivatives",
					"futures",
					"unified",
					"unknown"
				],
				"type": "string"
			},
...
			"ErrorResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"originalErrorMessage": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AuthParams": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"secretKey": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					},
					"exchangeId": {
						"type": "string"
					},
					"exchangeInstanceId": {
						"type": "string"
					}
				},
				"required": [
					"secretKey",
					"apiKey",
					"exchangeId",
					"exchangeInstanceId"
				],
				"type": "object"
			},
...
			"Record_WalletType.WalletType-Array_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ReturnType_CedeSDK-at-api_91_internalTransferRoutes_93__": {
				"$ref": "#/components/schemas/Record_WalletType.WalletType-Array_",
				"description": "Obtain the return type of a function type"
			},
			"InternalTransferRoutesResponse": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_internalTransferRoutes_93__"
			},
			"ReturnType_CedeSDK-at-api_91_internalTransfer_93__": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"description": "Obtain the return type of a function type"
			},
			"InternalTransferResponse": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_internalTransfer_93__"
			},
			"InternalTransferParams": {
				"properties": {
					"amount": {
						"type": "string"
					},
					"tokenSymbol": {
						"type": "string"
					},
					"fromWalletType": {
						"type": "string"
					},
					"toWalletType": {
						"type": "string"
					},
					"pairSymbolFrom": {
						"type": "string"
					},
					"pairSymbolTo": {
						"type": "string"
					},
					"auth": {
						"$ref": "#/components/schemas/AuthParams"
					}
				},
				"required": [
					"amount",
					"tokenSymbol",
					"fromWalletType",
					"toWalletType",
					"auth"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SubAccount": {
				"properties": {
					"uid": {
						"type": "string"
					}
				},
				"required": [
					"uid"
				],
				"type": "object"
			},
			"ReturnType_CedeSDK-at-api_91_getSubAccounts_93__": {
				"items": {
					"$ref": "#/components/schemas/SubAccount"
				},
				"type": "array",
				"description": "Obtain the return type of a function type"
			},
			"GetSubAccountsResponse": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getSubAccounts_93__"
			},
			"Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58_PortfolioBalanceEntry___": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ReturnType_CedeSDK-at-api_91_getSubAccountBalancesV2_93__": {
				"$ref": "#/components/schemas/Partial___91_walletType-string_93__58___91_tokenSymbol-string_93__58_PortfolioBalanceEntry___",
				"description": "Obtain the return type of a function type"
			},
			"GetSubAccountBalancesResponseV2": {
				"$ref": "#/components/schemas/ReturnType_CedeSDK-at-api_91_getSubAccountBalancesV2_93__"
			},
			"SubAccountTransferResponse": {
				"properties": {
					"transferId": {
						"type": "string"
					}
				},
				"required": [
					"transferId"
				],
				"type": "object"
			},
			"Pick_CedeSDKSubAccountTransferParams.Exclude_keyofCedeSDKSubAccountTransferParams.fromExchange-or-toExchange-or-readonlyExchange-or-exchange__": {
				"properties": {
					"tokenSymbol": {
						"type": "string"
					},
					"amount": {
						"type": "string"
					},
					"fromWalletType": {
						"$ref": "#/components/schemas/WalletType"
					},
					"toWalletType": {
						"$ref": "#/components/schemas/WalletType"
					},
					"fromUid": {
						"type": "string"
					},
					"toUid": {
						"type": "string"
					}
				},
				"required": [
					"tokenSymbol",
					"amount",
					"fromUid",
					"toUid"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_CedeSDKSubAccountTransferParams.fromExchange-or-toExchange-or-readonlyExchange-or-exchange_": {
				"$ref": "#/components/schemas/Pick_CedeSDKSubAccountTransferParams.Exclude_keyofCedeSDKSubAccountTransferParams.fromExchange-or-toExchange-or-readonlyExchange-or-exchange__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SubAccountTransferParams": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_CedeSDKSubAccountTransferParams.fromExchange-or-toExchange-or-readonlyExchange-or-exchange_"
					},
					{
						"properties": {
							"auth": {
								"$ref": "#/components/schemas/AuthParams"
							},
							"exchangeInstanceId": {
								"type": "string"
							}
						},
						"required": [
							"auth",
							"exchangeInstanceId"
						],
						"type": "object"
					}
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@cedelabs/sdk-http-api",
		"version": "1.2.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/transfer/routes": {
			"get": {
				"operationId": "GetInternalTransferRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalTransferRoutesResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get internal transfer routes.\nRetrieves all possible transfer routes between different wallet types.\nReturns a mapping of source wallet types to allowed destination wallet types.",
				"tags": [
					"Transfer"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/transfer": {
			"post": {
				"operationId": "InternalTransfer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalTransferResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Execute internal transfer.\nTransfers assets between different wallet types within the same exchange.\nSupports optional pair symbols for isolated margin transfers.",
				"tags": [
					"Transfer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InternalTransferParams"
							}
						}
					}
				}
			}
		},
...
		"/subAccount": {
			"get": {
				"operationId": "GetSubAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSubAccountsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all sub accounts of the master account. The response is an array of sub-account UIDs.",
				"tags": [
					"Sub Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/subAccount/balances": {
			"get": {
				"operationId": "GetSubAccountBalances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSubAccountBalancesResponseV2"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get sub account balances for a specific sub account UID.",
				"tags": [
					"Sub Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-instance-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-password",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "x-exchange-api-uid",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/subAccount/transfer": {
			"post": {
				"operationId": "SubAccountTransfer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SubAccountTransferResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"408": {
						"description": "Request Timeout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Service Unavailable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Transfer funds between sub-account and master account.",
				"tags": [
					"Sub Accounts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SubAccountTransferParams"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1"
		}
	]
}